+++ ledger/ledger-api-integration-tests/src/test/lib/scala/com/digitalasset/platform/apitesting/TestParties.scala
@@ -1,0 +1,18 @@
+// Copyright (c) 2019 The DAML Authors. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.digitalasset.platform.apitesting
+
+object TestParties {
+  val Alice = "Alice"
+  val Bob = "Bob"
+  val Charlie = "Charlie"
+  val Dan = "Dan"
+  val Eve = "Eve"
+  val Frank = "Frank"
+  val Grace = "Grace"
+  val Heidi = "Heidi"
+  val Ivan = "Ivan"
+
+  val AllParties = List(Alice, Bob, Charlie, Dan, Eve, Frank, Grace, Heidi, Ivan)
+}
